Pull rate limiting types into a separate file
